additional module refactoring to break the circular dependency room - ai server
