+++ src/domain/access/invitation/invitation.module.ts
@@ -13,6 +13,7 @@ import { InvitationLifecycleResolverFields } from './invitation.resolver.fields.
 import { InvitationLifecycleService } from './invitation.service.lifecycle';
 import { AccountLookupModule } from '@domain/space/account.lookup/account.lookup.module';
 import { VirtualContributorLookupModule } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.module';
+import { UserLookupModule } from '@domain/community/user-lookup/user.lookup.module';
 
 @Module({
   imports: [
@@ -21,6 +22,7 @@ import { VirtualContributorLookupModule } from '@domain/community/virtual-contri
     LifecycleModule,
     ContributorModule,
     VirtualContributorLookupModule,
+    UserLookupModule,
     AccountLookupModule,
     TypeOrmModule.forFeature([Invitation]),
   ],
