+++ src/domain/communication/room/room.resolver.mutations.ts
@@ -27,11 +27,11 @@ import { IMessageReaction } from '../message.reaction/message.reaction.interface
 import { SubscriptionPublishService } from '@services/subscriptions/subscription-service';
 import { MutationType } from '@common/enums/subscriptions';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
-import { RoomServiceMentions } from './room.service.mentions';
 import { Mention } from '../messaging/mention.interface';
 import { IRoom } from './room.interface';
-import { VirtualContributorService } from '@domain/community/virtual-contributor/virtual.contributor.service';
 import { VirtualContributorMessageService } from '../virtual.contributor.message/virtual.contributor.message.service';
+import { VirtualContributorLookupService } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.service';
+import { RoomMentionsService } from '../room-mentions/room.mentions.service';
 
 @Resolver()
 export class RoomResolverMutations {
@@ -41,10 +41,10 @@ export class RoomResolverMutations {
     private namingService: NamingService,
     private roomAuthorizationService: RoomAuthorizationService,
     private roomServiceEvents: RoomServiceEvents,
-    private roomServiceMentions: RoomServiceMentions,
+    private roomMentionsService: RoomMentionsService,
     private subscriptionPublishService: SubscriptionPublishService,
     private virtualContributorMessageService: VirtualContributorMessageService,
-    private virtualContributorsService: VirtualContributorService,
+    private virtualContributorLookupService: VirtualContributorLookupService,
     @Inject(WINSTON_MODULE_NEST_PROVIDER) private readonly logger: LoggerService
   ) {}
 
@@ -74,7 +74,7 @@ export class RoomResolverMutations {
 
     const accessVirtualContributors = this.virtualContributorsEnabled();
 
-    const mentions = await this.roomServiceMentions.getMentionsFromText(
+    const mentions = await this.roomMentionsService.getMentionsFromText(
       messageData.message
     );
 
@@ -91,7 +91,7 @@ export class RoomResolverMutations {
           messageData.roomID
         );
 
-        this.roomServiceMentions.processNotificationMentions(
+        this.roomMentionsService.processNotificationMentions(
           mentions,
           post.id,
           post.nameID,
@@ -116,7 +116,7 @@ export class RoomResolverMutations {
           agentInfo
         );
         if (accessVirtualContributors) {
-          await this.roomServiceMentions.processVirtualContributorMentions(
+          await this.roomMentionsService.processVirtualContributorMentions(
             mentions,
             message.message,
             threadID,
@@ -131,7 +131,7 @@ export class RoomResolverMutations {
           messageData.roomID
         );
 
-        this.roomServiceMentions.processNotificationMentions(
+        this.roomMentionsService.processNotificationMentions(
           mentions,
           calendar.id,
           calendar.nameID,
@@ -147,7 +147,7 @@ export class RoomResolverMutations {
           messageData.roomID
         );
 
-        this.roomServiceMentions.processNotificationMentions(
+        this.roomMentionsService.processNotificationMentions(
           mentions,
           discussion.id,
           discussion.nameID,
@@ -161,7 +161,7 @@ export class RoomResolverMutations {
         const discussionForum = await this.namingService.getDiscussionForRoom(
           messageData.roomID
         );
-        this.roomServiceMentions.processNotificationMentions(
+        this.roomMentionsService.processNotificationMentions(
           mentions,
           discussionForum.id,
           discussionForum.nameID,
@@ -191,7 +191,7 @@ export class RoomResolverMutations {
         );
 
         // Mentions notifications should be sent regardless of callout visibility per client-web#5557
-        this.roomServiceMentions.processNotificationMentions(
+        this.roomMentionsService.processNotificationMentions(
           mentions,
           callout.id,
           callout.nameID,
@@ -202,7 +202,7 @@ export class RoomResolverMutations {
         );
 
         if (accessVirtualContributors) {
-          await this.roomServiceMentions.processVirtualContributorMentions(
+          await this.roomMentionsService.processVirtualContributorMentions(
             mentions,
             message.message,
             threadID,
@@ -275,7 +275,7 @@ export class RoomResolverMutations {
 
     const accessVirtualContributors = this.virtualContributorsEnabled();
     const mentions: Mention[] =
-      await this.roomServiceMentions.getMentionsFromText(messageData.message);
+      await this.roomMentionsService.getMentionsFromText(messageData.message);
     const threadID = messageData.threadID;
 
     const messageOwnerId = await this.roomService.getUserIdForMessage(
@@ -321,12 +321,13 @@ export class RoomResolverMutations {
         // TODO extact in a helper function
         if (accessVirtualContributors) {
           // Check before processing so as not to reply to same message where interaction started
-          const vcInteraction = await this.roomService.getVcInteractionByThread(
-            room.id,
-            threadID
-          );
+          const vcInteraction =
+            await this.roomMentionsService.getVcInteractionByThread(
+              room.id,
+              threadID
+            );
 
-          await this.roomServiceMentions.processVirtualContributorMentions(
+          await this.roomMentionsService.processVirtualContributorMentions(
             mentions,
             messageData.message,
             threadID,
@@ -340,10 +341,10 @@ export class RoomResolverMutations {
               LogContext.VIRTUAL_CONTRIBUTOR
             );
             const contextSpaceID =
-              await this.roomServiceMentions.getSpaceIdForRoom(room);
+              await this.roomMentionsService.getSpaceIdForRoom(room);
 
             const vcMentioned =
-              await this.virtualContributorsService.getVirtualContributorOrFail(
+              await this.virtualContributorLookupService.getVirtualContributorOrFail(
                 vcInteraction.virtualContributorID
               );
 
@@ -414,11 +415,12 @@ export class RoomResolverMutations {
 
         if (accessVirtualContributors) {
           // Check before processing so as not to reply to same message where interaction started
-          const vcInteraction = await this.roomService.getVcInteractionByThread(
-            room.id,
-            threadID
-          );
-          await this.roomServiceMentions.processVirtualContributorMentions(
+          const vcInteraction =
+            await this.roomMentionsService.getVcInteractionByThread(
+              room.id,
+              threadID
+            );
+          await this.roomMentionsService.processVirtualContributorMentions(
             mentions,
             messageData.message,
             threadID,
@@ -432,11 +434,11 @@ export class RoomResolverMutations {
               LogContext.VIRTUAL_CONTRIBUTOR
             );
             const vcMentioned =
-              await this.virtualContributorsService.getVirtualContributorOrFail(
+              await this.virtualContributorLookupService.getVirtualContributorOrFail(
                 vcInteraction.virtualContributorID
               );
             const contextSpaceID =
-              await this.roomServiceMentions.getSpaceIdForRoom(room);
+              await this.roomMentionsService.getSpaceIdForRoom(room);
 
             await this.virtualContributorMessageService.invokeVirtualContributor(
               vcMentioned?.nameID,
@@ -465,7 +467,7 @@ export class RoomResolverMutations {
             agentInfo,
             messageOwnerId
           );
-          this.roomServiceMentions.processNotificationMentions(
+          this.roomMentionsService.processNotificationMentions(
             mentions,
             callout.id,
             callout.nameID,
