+++ src/domain/communication/virtual.contributor.message/virtual.contributor.message.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { VirtualContributorMessageService } from './virtual.contributor.message.service';
-import { ContributorLookupModule } from '@services/infrastructure/contributor-lookup/contributor.lookup.module';
 import { AiServerAdapterModule } from '@services/adapters/ai-server-adapter/ai.server.adapter.module';
 import { VcInteractionModule } from '../vc-interaction/vc.interaction.module';
+import { VirtualContributorLookupModule } from '@domain/community/virtual-contributor-lookup/virtual.contributor.lookup.module';
 
 @Module({
   imports: [
-    ContributorLookupModule,
+    VirtualContributorLookupModule,
     AiServerAdapterModule,
     VcInteractionModule,
   ],
