+++ src/domain/innovation-hub/innovation.hub.module.ts
@@ -11,11 +11,11 @@ import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { StorageAggregatorResolverModule } from '@services/infrastructure/storage-aggregator-resolver/storage.aggregator.resolver.module';
 import { InnovationHubResolverFields } from './innovation.hub.resolver.fields';
-import { AccountHostModule } from '@domain/space/account.host/account.host.module';
+import { AccountLookupModule } from '@domain/space/account.lookup/account.lookup.module';
 
 @Module({
   imports: [
-    AccountHostModule,
+    AccountLookupModule,
     SpaceModule,
     ProfileModule,
     AuthorizationPolicyModule,
