+++ src/platform/admin/licensing/admin.licensing.module.ts
@@ -5,14 +5,14 @@ import { AdminLicensingService } from './admin.licensing.service';
 import { LicenseIssuerModule } from '@platform/licensing/credential-based/license-credential-issuer/license.issuer.module';
 import { SpaceModule } from '@domain/space/space/space.module';
 import { AccountModule } from '@domain/space/account/account.module';
-import { AccountHostModule } from '@domain/space/account.host/account.host.module';
 import { LicensingFrameworkModule } from '@platform/licensing/credential-based/licensing-framework/licensing.framework.module';
 import { LicenseModule } from '@domain/common/license/license.module';
+import { AccountLookupModule } from '@domain/space/account.lookup/account.lookup.module';
 
 @Module({
   imports: [
     AccountModule,
-    AccountHostModule,
+    AccountLookupModule,
     SpaceModule,
     LicensingFrameworkModule,
     LicenseModule,
